+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+* Add paginated variant of `Users.lists`. The original method is deprecated.
+
 ## 6.20.1 - 2026-06-18
 
 * When parsing a `Rating` with an unknown value, return `null` instead of throwing `ArrayIndexOutOfBoundsException`. Background: the Trakt API was seen returning `0` as a rating value.